airsim/requirements.txt
airsim==1.6.0
anyio==3.3.2
argon2-cffi==21.1.0
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
cuda==0.0.1
cupy-cuda112==9.1.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
djitellopy==2.4.0
entrypoints==0.3
fastrlock==0.6
gym==0.18.3
idna==2.10
imageio==2.9.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.4
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-packaging==0.10.6
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
line-profiler==3.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==4.4.1
numpy==1.20.2
open3d==0.13.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.3.56
packaging==21.0
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.2
path==15.1.2
perlin-numpy @ git+https://github.com/pvigier/perlin-numpy@6f077f811f5708e504732c26dee8f2015b95da0c
pickleshare==0.7.5
Pillow==8.3.2
plotly==4.14.3
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pycodestyle==2.7.0
pycparser==2.20
pyglet==1.5.15
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
requests-unixsocket==0.2.0
retrying==1.3.3
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.2
Send2Trash==1.8.0
six==1.16.0
snakeviz==2.1.0
sniffio==1.2.0
soupsieve==2.2.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.3.31
toml==0.10.2
tomlkit==0.7.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==4.5.3
tqdm==4.61.0
traitlets==5.1.0
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.1
wincertstore==0.2
